requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
aiohttp>=3.8.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.6.0,<3.0.0
websocket-client>=1.4.0
pandas>=2.0.0,<3.0.0
pyarrow>=14.0.0,<20.0.0
